The message sender is automatically settled how session fan. There is now a checking if the fan that want erase the message is the sender of it. Integration tests were created to check the authorization of operations automatically.

+++ src/main/java/com/furia/chat/service/ChatMessageService.java
@@ -4,12 +4,11 @@ import com.furia.chat.dto.ChatMessageRequest;
 import com.furia.chat.model.ChatMessage;
 import com.furia.chat.repository.ChatMessageRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
-
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ChatMessageService {
@@ -20,8 +19,12 @@ public class ChatMessageService {
         if (message.message() == null || message.message().trim().isEmpty()) {
             throw new IllegalArgumentException("Message should be not empty");
         }
-
-        ChatMessage filled = new ChatMessage(message.sender(), message.message());
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String sender = auth.getName();
+        if (auth.getAuthorities().iterator().next().getAuthority().equals("ROLE_ADMIN")) {
+            sender = "FURIA";
+        }
+        ChatMessage filled = new ChatMessage(sender, message.message());
         return chatMessageRepository.save(filled);
     }
 
@@ -29,9 +32,15 @@ public class ChatMessageService {
         return chatMessageRepository.findAll();
     }
 
-    public void deleteById(String id) throws ClassNotFoundException {
+    public void deleteById(String id) throws ClassNotFoundException, AccessDeniedException {
         ChatMessage chatMessage = chatMessageRepository.findById(id)
-                .orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND, "Message not found"));
+                .orElseThrow(() -> new ClassNotFoundException("Not found"));
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String sender = auth.getName();
+        String role = auth.getAuthorities().iterator().next().getAuthority();
+        if (!chatMessage.getSender().equals(sender) && !role.equals("ROLE_ADMIN")) {
+            throw new AccessDeniedException("No permission to delete this message");
+        }
         chatMessageRepository.delete(chatMessage);
     }
 }
